Move logging options to schema generator.
