+++ src/Tests/Microsoft.Diagnostics.Monitoring.ConfigurationSchema/Options/ConsoleFormatterOptions.cs
@@ -3,19 +3,13 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.Monitoring.WebApi;
-using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Text.Json;
+using System.ComponentModel;
 
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
+namespace Microsoft.Diagnostics.Monitoring.Options
 {
-    internal sealed class JsonConsoleFormatterOptions
+    internal sealed class ConsoleFormatterOptions
     {
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_JsonConsoleFormatterOptions_WriterOptions))]
-        public JsonWriterOptions JsonWriterOptions { get; set; }
-
         [Display(
             ResourceType = typeof(OptionsDisplayStrings),
             Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_ConsoleFormatterOptions_IncludeScopes))]
