+++ src/Tests/Microsoft.Diagnostics.Monitoring.ConfigurationSchema/Options/ConsoleLoggerOptions.cs
@@ -8,7 +8,7 @@ using System.ComponentModel;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
+namespace Microsoft.Diagnostics.Monitoring.Options
 {
     internal sealed class ConsoleLoggerOptions
     {
