+++ src/Tests/Microsoft.Diagnostics.Monitoring.ConfigurationSchema/Options/LoggingOptions.cs
@@ -7,9 +7,8 @@ using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
 
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
+namespace Microsoft.Diagnostics.Monitoring.Options
 {
     internal sealed class LoggingOptions
     {
