+++ src/Tests/Microsoft.Diagnostics.Monitoring.ConfigurationSchema/Options/SimpleConsoleFormatterOptions.cs
@@ -4,15 +4,11 @@
 
 using Microsoft.Diagnostics.Monitoring.WebApi;
 using Microsoft.Extensions.Logging.Console;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
-using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
+namespace Microsoft.Diagnostics.Monitoring.Options
 {
     internal sealed class SimpleConsoleFormatterOptions
     {
